refactor(): continued split ups & cleanups
