+++ src/core/geometry/poly3/create.js
@@ -53,6 +53,7 @@ function create () {
   return {
     vertices: [],
     shared: {
+      color: undefined,
       getTag: function () {
         let result = this.tag
         if (!result) {
@@ -60,11 +61,6 @@ function create () {
           this.tag = result
         }
         return result
-      },
-      // get a string uniquely identifying this object
-      getHash: function () {
-        if (!this.color) return 'null'
-        return this.color.join('/')
       }
     },
     plane: [0, 0, 0, 1] // FIXME: use plane structure, vec4, plane class?
