+++ src/core/geometry/shape2/index.js
@@ -22,6 +22,9 @@ module.exports = {
   toCompactBinary: require('./toCompactBinary'),
   toCSGWall: require('./toCSGWall'),
   toPoints: require('./toPoints'),
+  toVec3Pairs: require('./toVec3Pairs'),
+  toPlanePolygons: require('./toPlanePolygons'),
+  toWallPolygons: require('./toWallPolygons'),
   toString: require('./toString'),
   subtract: require('./subtract'),
   transform: require('./transform'),
